+++ cursor-shape.h
@@ -26,4 +26,4 @@ const char *const *cursor_shape_to_string(enum cursor_shape shape);
 enum wp_cursor_shape_device_v1_shape cursor_shape_to_server_shape(
     enum cursor_shape shape);
 enum wp_cursor_shape_device_v1_shape cursor_string_to_server_shape(
-    const char *xcursor);
+    const char *xcursor, int bound_version);